requirements.txt
astroid==2.3.3
Babel==2.7.0
colorama==0.4.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.7
django-bootstrap3==11.1.0
django-phonenumber-field==3.0.1
gunicorn==20.0.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
phonenumbers==8.10.23
Pillow==6.2.1
psycopg2==2.8.4
pylint==2.4.4
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
static3==0.7.0
typed-ast==1.4.0
wrapt==1.11.2
